Prevent deadlocks in dispatching
Changed the queue type to something that can really queue up stuff. The
previous implementation used a SynchronousQueue, which doesn't provide a
real thread decoupling because there is only one slot available for
events to dispatch. As soon as this slot is filled the whole internal
connector logic is blocked by the user level waiting until the event is
really dispatched. This already lead to a deadlock in the integration
test.