Move introspection enabled information to participant config
Allow to chose whether introspection is enabled for each participant by
moving this information to the participant config. As a consequence, the
participant observer for the introspection can now be registered all the
time and recursion is prevented by disabling introspection for the
internal introspection participants.

Moreover, this commit ensures that the introspection is disabled
correctly with the last client participant being destructed.