Introspection created events from: construction -> activate calls
This commit transfers emitting introspection created events from
participant constructions (via the factory) to the time the activate
method is called. This prevents potential asymmetries in case a
participant is created but not activated. In that case, such a
participant could probably not be deactivated and hence the
introspection never noticed that is was destroyed.

In order to fulfill this, some parts of the (more or less) internal API
had to be changed. All Participant implementations now only have
constructors which accepts ParticipantCreateArgs.