Handle read-only restrictions of ByteBuffers correctly
It is invalid to call array() on read-only ByteBuffer instances. This
was done in converters like the DoubleConverter to access the wire data
for deserialization. However, in case of the socket transport when
handling send-receive cycles in the same process, the ByteBuffer that
ended up at the receiving side for deserialization was such a read only
instance.

This commit works around this issue by using non-read-only instances all
the time. Ultimately, this is still a hacky solution because ByteBuffers
with their statefull reading interface are actually a bad choice for
Converters, because multiple converters might read them in parallel.

fixes #2181

* ConnectorCheck: new test case for this issue
* SocketInPushConnector: always create writable ByteBuffer instances