Fix handshake-activation race conditions in socket transport
Fix several possible race conditions where still inactive connections
could end up in the connection list of a BusServer instance and a
potential sending in this state would lead to an exception. This is
achieved by making the handshake part of the activation process so that
locks are held a bit longer and by accepting that a connection might
still be inactive in case a notification has to be sent on it. This case
has to occur in the BusServer, because we also want to ensure that
BusClientConnections can only return from activate once they are
perfectly integrated in the server dispatching system. If the handshake
was completed before integrating them in the connection list of the
BusServer, there is a time gap where events might not be dispatched to
these connections.