Use the default Future interface for the client API
Instead of our custom Future implementation, expose the
java.util.concurrent.Future interface in the client API. This ensures
compatibility with standard interfaces, prevents accidental exception
hiding (as we had before with interrupted exceptions) and prevents
clients from completing a Future instance.

This changes the API slightly, since the get method with simply a long
value for a millisecond timeout does not exist in the standard
interface. However, the existing versions with a TimeUnit are much
clearer to read anyway.

* examples/ClientExample.java: adapt to changed future, comment about
  time unit is now redundant, remote
* src/rsb/patterns/Future.java: make package private
* src/rsb/patterns/RemoteServer.java
  Use standard future for external interface
* test/rsb/patterns/RpcIntegrationTest.java: adapt to changed interface