Upgrade to latest PMD plugin
Upgrade to the latest PMD maven plugin and with that to PMD version
6.8.0. This required a rewrite of the rules file due to a changed rule
structure in PMD. The ruleset as been rebuilt based on the old
configuration and some new rules that seem reasonable have been added.
The source code has been adapted to get back to the zero warnings
policy.